requirements.txt
requests
beautifulsoup4
python-socketio
checklib
